予想/ai-auto-mailer/boatrace-collector/requirements.txt
pyjpboatrace>=0.5.0
psycopg2-binary>=2.9.0
requests>=2.28.0
lhafile>=0.3.0
beautifulsoup4>=4.12.0
apscheduler>=3.10.0
pymysql>=1.1.0
